+++ packages/shared/src/index.ts
@@ -2,3 +2,4 @@
 export * from './types/index.ts'
 export * from './parser/index.ts'
 export * from './linter/index.ts'
+export * from './yaml-dump-opts.ts'